Add first integration with auth0 sdk
