+++ .env.development
@@ -1,1 +1,6 @@
 BACK_ENDPOINT="https://evercode-backend-staging.herokuapp.com"
+AUTH0_SECRET='dshaduiehciasc3y4yu3gf7fgddvc7g3fefcudbfuyewfhbajchdvhdsfhj'
+AUTH0_BASE_URL='http://localhost:3000'
+AUTH0_ISSUER_BASE_URL='https://dev-yvx59wsh.us.auth0.com'
+AUTH0_CLIENT_ID='46MkQUYqSlagNp7evBVomuYRBqSgcsAl'
+AUTH0_CLIENT_SECRET='qZz6eYtFselN8rLOPZlI5h7LdLqXwyeSEvm7wZJ4N5-s6S03GHERFS1mvNRzi41z'
