+++ package.json
@@ -12,6 +12,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@auth0/nextjs-auth0": "^1.7.0",
     "axios": "^0.26.1",
     "next": "12.1.4",
     "react": "18.0.0",
